pyproject.toml
[tool.poetry]
name = "autondance-demo-server"
version = "0.1.0"
description = "Demo server for autodance"
authors = ["Praveen Perumal <ppraveen98841@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
aiofiles = "^23.2.1"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
